KYC_VERIFICATION/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic[email]==2.9.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.1

# HTTP client for real third-party API calls (Karza, Cashfree, DigiLocker)
requests==2.32.3

# Dummy data generation
faker==24.0.0